✨ Add sitemap.xml
Having a sitemap to all of our sites will allow us to statically
generate across all the different content repositories that we have.

This is a simple site, with simple content, so all that is needed is
to add `sitemap.xml.ts` route that walks across the document table of
contents stored in `toc.json` and create the manifest of urls
contained therein.