Make this site embeddable into another
Because of the way fresh works it uses "clean" urls, relative paths
don't really work.

This gets rid of all "basing" elements and uses absolute urls which
are relative to the root of this site. Since we're rebasing on the
consuming edge, this is fine.

In the future, we'll use directory urls for everything.